auth_service1/auth-back/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
alembic==1.11.1
pydantic==2.11.1
email-validator==2.0.0.post2